\name{BacktastingTools-package}
\alias{BacktastingTools-package}
\alias{BacktastingTools}
\docType{package}
\title{
\packageTitle{BacktastingTools}
}
\description{
\packageDescription{BacktastingTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BacktastingTools}
\packageIndices{BacktastingTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BacktastingTools}

Maintainer: \packageMaintainer{BacktastingTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
